+++ src/main/java/cl/transbank/pos/exceptions/integrado/TransbankDetailException.java
@@ -1,7 +1,7 @@
-package cl.transbank.pos.exceptions.integradoExceptions;
+package cl.transbank.pos.exceptions.integrado;
 
-import cl.transbank.pos.exceptions.commonExceptions.TransbankException;
-import cl.transbank.pos.responses.integradoResponses.DetailResponse;
+import cl.transbank.pos.exceptions.common.TransbankException;
+import cl.transbank.pos.responses.integrado.DetailResponse;
 
 public class TransbankDetailException extends TransbankException {
     private DetailResponse detailResponse;
