+++ src/main/java/cl/transbank/pos/exceptions/integrado/TransbankLastSaleException.java
@@ -1,7 +1,7 @@
-package cl.transbank.pos.exceptions.integradoExceptions;
+package cl.transbank.pos.exceptions.integrado;
 
-import cl.transbank.pos.exceptions.commonExceptions.TransbankException;
-import cl.transbank.pos.responses.integradoResponses.LastSaleResponse;
+import cl.transbank.pos.exceptions.common.TransbankException;
+import cl.transbank.pos.responses.integrado.LastSaleResponse;
 
 public class TransbankLastSaleException extends TransbankException {
     private LastSaleResponse saleResponse;
