+++ src/main/java/cl/transbank/pos/exceptions/integrado/TransbankMultiCodeDetailException.java
@@ -1,4 +1,4 @@
-package cl.transbank.pos.exceptions.integradoExceptions;
+package cl.transbank.pos.exceptions.integrado;
 
 public class TransbankMultiCodeDetailException extends TransbankDetailException {
     public TransbankMultiCodeDetailException(String message, Throwable cause) {super(message, cause);}
