+++ src/main/java/cl/transbank/pos/exceptions/integrado/TransbankSaleException.java
@@ -1,7 +1,7 @@
-package cl.transbank.pos.exceptions.integradoExceptions;
+package cl.transbank.pos.exceptions.integrado;
 
-import cl.transbank.pos.exceptions.commonExceptions.TransbankException;
-import cl.transbank.pos.responses.integradoResponses.SaleResponse;
+import cl.transbank.pos.exceptions.common.TransbankException;
+import cl.transbank.pos.responses.integrado.SaleResponse;
 
 public class TransbankSaleException extends TransbankException {
     private SaleResponse saleResponse;
