+++ src/main/java/cl/transbank/pos/exceptions/integrado/TransbankTotalsException.java
@@ -1,7 +1,7 @@
-package cl.transbank.pos.exceptions.integradoExceptions;
+package cl.transbank.pos.exceptions.integrado;
 
-import cl.transbank.pos.exceptions.commonExceptions.TransbankException;
-import cl.transbank.pos.responses.integradoResponses.TotalsResponse;
+import cl.transbank.pos.exceptions.common.TransbankException;
+import cl.transbank.pos.responses.integrado.TotalsResponse;
 
 public class TransbankTotalsException extends TransbankException {
     private TotalsResponse totalsResponse;
